requirements.txt
eventlet==0.25.0
Flask_SocketIO==4.2.1
numpy==1.17.2
Flask==1.1.1
tensorflow_gpu==1.14.0
opencv_python==3.4.4.19
Pillow==6.2.1
object_detection==0.0.3
tensorflow==2.0.0
